Changed addsubject, and added custaddition
addsubject --> Changed it back to the original. BUT instead of looking
for a character the 'if' statement looks for an 'int'. anything else
will have the word 'output' added to it.
ALSO: the addsubject file is working with the old_input.dat file!!!

custaddition --> This new program is a program that allows you to
specify a start and end point (in the arguments respectively) and turns
that ranged of output.dat files into output.mat files :D