changed addsubject and created multidisp
addsubject --> added an additional if - elseif statment that works with
the input the user gave. if the user gave a string then the function
will have the function getSubjectParams search for that string. IF the
user gives something else (an integer) then the function turns that
integer into a string and has the function getSubjectParams search for
that particular subject number instead
ALSO: commented code out at the bottom just before the function saves
the information into a .mat file

mulidisp --> this is a function that will compare the errors to the
trial number of baseline, extraction, and washout phases and will stack
them for all 8 patients